+++ server/.settings/org.eclipse.wst.ws.service.policy.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+org.eclipse.wst.ws.service.policy.projectEnabled=false
