+++ server/src/db/MongoDBConnection.java
@@ -1,0 +1,203 @@
+package db;
+
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import model.Restaurant;
+
+import org.bson.Document;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import yelp.YelpAPI;
+
+import com.mongodb.Block;
+import com.mongodb.MongoClient;
+import com.mongodb.client.FindIterable;
+import com.mongodb.client.MongoDatabase;
+
+import static com.mongodb.client.model.Filters.eq;
+import static com.mongodb.client.model.Filters.regex;
+
+public class MongoDBConnection implements DBConnection {
+
+	private static final int MAX_RECOMMENDED_RESTAURANTS = 10;
+
+	private MongoClient mongoClient;
+	private MongoDatabase db;
+
+	public MongoDBConnection() {
+		// Connects to local mongodb server.
+		mongoClient = new MongoClient();
+		db = mongoClient.getDatabase(DBUtil.DB_NAME);
+
+	}
+
+	@Override
+	public void close() {
+		if (mongoClient != null) {
+			mongoClient.close();
+		}
+	}
+
+	@Override
+	public void setVisitedRestaurants(String userId, List<String> businessIds) {
+		db.getCollection("users").updateOne(new Document("user_id", userId),
+				new Document("$pushAll", new Document("visited", businessIds)));
+
+	}
+
+	@Override
+	public void unsetVisitedRestaurants(String userId, List<String> businessIds) {
+		db.getCollection("users").updateOne(new Document("user_id", userId),
+				new Document("$pullAll", new Document("visited", businessIds)));
+	}
+
+	@Override
+	public Set<String> getVisitedRestaurants(String userId) {
+		Set<String> set = new HashSet<>();
+		FindIterable<Document> iterable = db.getCollection("users").find(
+				new Document("user_id", userId));
+
+		iterable.forEach(new Block<Document>() {
+			@Override
+			public void apply(final Document document) {
+				List<String> list = (List<String>) document.get("visited");
+				set.addAll(list);
+			}
+		});
+		return set;
+	}
+
+	@Override
+	public JSONObject getRestaurantsById(String businessId, boolean isVisited) {
+
+		FindIterable<Document> iterable = db.getCollection("restaurants").find(
+				eq("business_id", businessId));
+		try {
+			return new JSONObject(iterable.first().toJson());
+		} catch (JSONException e) {
+			e.printStackTrace();
+		}
+		return null;
+	}
+
+	@Override
+	public JSONArray recommendRestaurants(String userId) {
+		try {
+
+			Set<String> visitedRestaurants = getVisitedRestaurants(userId);
+			Set<String> allCategories = new HashSet<>();// why hashSet?
+			for (String restaurant : visitedRestaurants) {
+				allCategories.addAll(getCategories(restaurant));
+			}
+			Set<String> allRestaurants = new HashSet<>();
+			for (String category : allCategories) {
+				Set<String> set = getBusinessId(category);
+				allRestaurants.addAll(set);
+			}
+			Set<JSONObject> diff = new HashSet<>();
+			int count = 0;
+			for (String businessId : allRestaurants) {
+				// Perform filtering
+				if (!visitedRestaurants.contains(businessId)) {
+					diff.add(getRestaurantsById(businessId, false));
+					count++;
+					if (count >= MAX_RECOMMENDED_RESTAURANTS) {
+						break;
+					}
+				}
+			}
+			return new JSONArray(diff);
+		} catch (Exception e) {
+			System.out.println(e.getMessage());
+		}
+		return null;
+	}
+
+	@Override
+	public JSONArray searchRestaurants(String userId, double lat, double lon) {
+		try {
+			YelpAPI api = new YelpAPI();
+			JSONObject response = new JSONObject(
+					api.searchForBusinessesByLocation(lat, lon));
+			JSONArray array = (JSONArray) response.get("businesses");
+
+			List<JSONObject> list = new ArrayList<JSONObject>();
+			Set<String> visited = getVisitedRestaurants(userId);
+
+			for (int i = 0; i < array.length(); i++) {
+				JSONObject object = array.getJSONObject(i);
+				Restaurant restaurant = new Restaurant(object);
+				String businessId = restaurant.getBusinessId();
+				String name = restaurant.getName();
+				String categories = restaurant.getCategories();
+				String city = restaurant.getCity();
+				String state = restaurant.getState();
+				String fullAddress = restaurant.getFullAddress();
+				double stars = restaurant.getStars();
+				double latitude = restaurant.getLatitude();
+				double longitude = restaurant.getLongitude();
+				String imageUrl = restaurant.getImageUrl();
+				String url = restaurant.getUrl();
+				JSONObject obj = restaurant.toJSONObject();
+				if (visited.contains(businessId)) {
+					obj.put("is_visited", true);
+				} else {
+					obj.put("is_visited", false);
+				}
+				db.getCollection("restaurants").insertOne(
+						new Document().append("business_id", businessId)
+								.append("name", name)
+								.append("categories", categories)
+								.append("city", city).append("state", state)
+								.append("fulle_address", fullAddress)
+								.append("stars", stars)
+								.append("latitude", latitude)
+								.append("longitude", longitude)
+								.append("image_url", imageUrl)
+								.append("url", url));
+				list.add(obj);
+			}
+			return new JSONArray(list);
+		} catch (Exception e) {
+			System.out.println(e.getMessage());
+		}
+		return null;
+	}
+
+	@Override
+	public Set<String> getCategories(String businessId) {
+		Set<String> set = new HashSet<>();
+		FindIterable<Document> iterable = db.getCollection("restaurants").find(
+				eq("business_id", businessId));
+		iterable.forEach(new Block<Document>() {
+			@Override
+			public void apply(final Document document) {
+				String[] categories = document.getString("categories").split(
+						",");
+				for (String category : categories) {
+					set.add(category.trim());
+				}
+			}
+		});
+		return set;
+	}
+
+	@Override
+	public Set<String> getBusinessId(String category) {
+		Set<String> set = new HashSet<>();
+		FindIterable<Document> iterable = db.getCollection("restaurants").find(
+				regex("categories", category));
+		iterable.forEach(new Block<Document>() {
+			@Override
+			public void apply(final Document document) {
+				set.add(document.getString("business_id"));
+			}
+		});
+		return set;
+	}
+}
