+++ server/src/db/ReviewImport.java
@@ -12,7 +12,7 @@ public class ReviewImport {
 	static {
 		try {
 			Class.forName("com.mysql.jdbc.Driver").newInstance();
-			conn = DriverManager.getConnection(DBUtil.URL);
+			conn = DriverManager.getConnection(DBUtil.MYSQL_URL);
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
